[release test] move azure related functions to `cloud_util.py`
out of `util.py`. also adding its own `py_library`
