+++ release/ray_release/file_manager/job_file_manager.py
@@ -9,6 +9,7 @@ from azure.identity import DefaultAzureCredential
 import boto3
 from google.cloud import storage
 from ray_release.aws import RELEASE_AWS_BUCKET
+from ray_release.cloud_util import generate_tmp_cloud_storage_path
 from ray_release.cluster_manager.cluster_manager import ClusterManager
 from ray_release.exception import FileDownloadError, FileUploadError
 from ray_release.file_manager.file_manager import FileManager
@@ -16,7 +17,6 @@ from ray_release.job_manager import JobManager
 from ray_release.logger import logger
 from ray_release.util import (
     exponential_backoff_retry,
-    generate_tmp_cloud_storage_path,
     S3_CLOUD_STORAGE,
     GS_CLOUD_STORAGE,
     GS_BUCKET,
@@ -64,8 +64,7 @@ class JobFileManager(FileManager):
         )
 
     def _generate_tmp_cloud_storage_path(self):
-        location = f"tmp/{generate_tmp_cloud_storage_path()}"
-        return location
+        return f"tmp/{generate_tmp_cloud_storage_path()}"
 
     def download_from_cloud(
         self, key: str, target: str, delete_after_download: bool = False
